+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `listTenants` function in alloy service.
+
 ### Changed
 
+- Add list of tenants in alloy config methods's parameters.
 - Fine-tune alloy-metrics resource usage configuration to avoid causing issues for customer workload and cluster tests.
 
 ## [0.20.0] - 2025-03-18